+++ react/components/withSearchContext.tsx
@@ -1,7 +1,7 @@
 import React, { ComponentType } from 'react'
 import { SearchPageContext } from 'vtex.search-page-context/SearchPageContext'
 
-function withSearchContext(WrappedComponent: ComponentType) {
+function withSearchContext(WrappedComponent: ComponentType<any>) {
   // eslint-disable-next-line react/display-name
   return function(props: any) {
     return (